backend/requirements.txt
gunicorn==20.0.4
django==4.1
djangorestframework==3.14
pillow==9.4.0
djoser==2.1.0
django-filter==22.1
python-dotenv==1.0.0
django-debug-toolbar==3.8.1
reportlab==3.6.12
psycopg2-binary==2.8.6
pydantic==1.10.6
django-cors-headers==3.14.0